requirements.txt
# Core dependencies
numpy>=1.19.0
Pillow>=8.0.0
lxml>=4.6.0
pyyaml>=5.4.0
regex>=2021.8.0

# Optional dependencies
# PDF processing
PyPDF2>=2.0.0
pdfplumber>=0.5.0

# OCR (optional)
pytesseract>=0.3.0
opencv-python>=4.5.0
requests>=2.25.0
aiohttp>=3.8.0

# Rendering (optional)
matplotlib>=3.3.0
cairosvg>=2.5.0

# Development
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0
flake8>=3.9.0
